add ability to put page headings into the toc
This allows most of the content buried under <details> tags to be
exposed by default, as the headings can be jumped to easily through
the toc instead. I disliked how so much of the content was hidden,
as it concealed just how much information was actually contained
in a single page. It also makes it take more effort to read, since
you had to click on every section you were interested in, rather
than merely scrolling.